Finished treatment and started segmentation modules
I made the treatment module as generic as I could. It does rely on the data being processed being fairly uniform in it's properties. The segmentation module has a graphical interface and visual but also allows individual signals and files to be selected to test the methods. The underlying structure is very similar to the treatment module.