Segmentation
Got the module for segmentation to work decently. It will plot the original data and the segmented. The segmentation method can be selected in a drop down. The segmentation module itself is an app, but the segmentation methods themselves can be m-files. These detect if a single numeric time series is being segmented for viewing or if all the files need to be segmented. All the files need to be segmented in order to do object-based segmenting. With numeric inputs any input can be used. The results are saved in the processed data and are viewable in the raw data review module.