Worked on segmentation of IMU data with turns
Implemented attitude calculations and and algorithm to find the turns. There is plotting code within the algorithm that I used to picture the method, and is currently commented out. This worked well on one object but it needs to be tested on more. Would need to modify the individual segmentation option to segment the correct data type instead of only the selected data type.