Worked on segmentation review
The individual segmentation will now work for individual time series, objects and files. If the segmentation method is not compatible with the selection it will output an error to the log and not perform it. The preview will now display the segmentation start and end frames. There are results for the segmentation that are used for this but they need to be renamed to be more general.