Worked on generalizing load functions and started a more complete configuration capability.
Removed the hardcoded files types for the load functions. Added code for it to pull the file extension and equipment type pairs from the load function names. This is almost working but the feval is not. Added an element to the config file that stores the path to these functions. Started a config tab in the app so these paths could potentially be updated in the app. Created a function to load png files.