Made changes related to the starting configuration and loading data from Pfizer
The configuration code errored when the path to the load files was no longer found and the options were empty. Added a condition to skip setting the load options if no load files were found.
CSV files from Actigraph and Pfizer can now be imported. There is also a histogram plot option in the raw data review. The Actigraph files do have strings, dates and numeric values but only the numeric values are displayed right now.