Worked on synchronizing different signals which lead to other changes.
Analysis_custom is a template module for custom analyses. This version is developed to synchronize different time series. The module itself is not an app but an m-file so more people can use it.

Started an outline of the User Manual.

Updated BAR_App so the analysis methods can be detected in the configuration tab and used the same as the processing methods. Added code to the export so raw data can be exported. The raw and res data do have different enough formats that they might not be able to be exported in the same file. They'll also have different signals/results that will lead to mismatched columns in the export table.

Updated load_txt_V3D to detect included sampling frequencies. Including them will be the responsibility of the user since V3D will not export them unless you tell it to.

Created process_Merge to merge two data sets together. This unidirectional at the moment and requires objects to be combined into files. It was more difficult to make it omnidirectional and multi-level than I was hoping.