Worked on custom code to synchronize different time series
Most of this work was to get code working that synchronizes different time series. Some of the load functions have updated comment sections. Wrote a load function to import mat files exported from Qyalisys QTM. Changed the V3D export so all signals have the same name, moved the V3D type field to the BAR App object field. Restarted the grouping code. It now runs as soon as the data is imported to create the groups. The module to process the groupings still needs work. Also worked on the merging module. This looks to be working nicely now. The results can be exported to be reviewed outside the app.