Worked towards publication.
I added a lot of comments to the code that would be released with the app at the time of it's first publication. A number of methods have been moved over to the production version of the app from the beta test files. I also completed the figures for Time Lag, FNN and RQA. The User Manual has been updated with all the file types available through the load functions and I worked on the text for the main application section.