Added version information.
Made sure all mlapp files had a version label and code in the startup function to make it the same as the BAR_App. Add another error code that will print the version number to the top of the log at startup. All the version numbers are pulled from the BAR_App label.