Worked on publication requirements
Modified the auto grouping code that runs when data is uploaded to act only on the groups of the file to fix a bug. Something was cause groups to be dropped inconsistently. Created a validation folder and a validation dataset to test the export output, groupings and merge modules.

Modified the Future Work section of the user manual to be called Community Guidelines. Added a section for installation instructions.