Worked on making the Grouping Module more adapatable.
The Grouping Module will not remove object groups with the same information within across groups in the same index. This is within columns in the table. Added a feature that allows the user to designate tags and remove them from signals and add them to group information. This make more signal names the same and puts their values on subsequent rows in the same column of the export.