Updated manual and configuration tab.
Changed figure numbers to reflect their section in the manual so it is easier to add new figures without changed many file names. Added a text field to the Configuration tab to help find the config file when new users start the app.
Added a new load function for APDM data.
Added a checkbox to the General Review Module so the plot updates automatically when a selection is changed.