Added log file, retooled dimensional data treatment.
Code was added to the main app to create and write log messages to a log file. A new one is created each time the app is run.
Modified the some of the primary analysis modules to treat dimensional data differently. It now treats them directly as column oriented data. Previously the signal name was manipulated for data with multiple dimensions. This required the data export for results data to be retooled.
Added code to load_txt_V3D that can remove numerous MEAN and STDDEV signals.