Added additional modules and modified existing ones.
Added a module to match signals and calculate cross-signal metrics. The only metric available right now performs statistical parameter matching.
Added a module to calculate lyapunov exponents.
Made some changes to the RQA module. These were mostly around how it handles already processed data.
Fixed a bug in DAMVi that should clear an error when running on a new computer. If the Log folder was not present it would error. Now it will create the folder if it doesn't already exist.
Created an eventing module and adapted a peak picking method for it.
Added a module to perform surrogation.