Changed analysis_QST
The figure will now save to a subfolder and will use the group information if a full file path is not in the group information.
Quality measures are now calculated up to the peak or the button press and ignore data after it that could give a false measure of poor quality.